+++ packages/card/src/components/StatisticCard/index.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { ConfigProvider } from 'antd';
-import ProCard from '../../Card';
-import type { ProCardProps } from '../../Card';
+import Card from '../Card';
+import type { CardProps } from '../../type';
 import classNames from 'classnames';
 import Statistic from '../Statistic';
 import type { StatisticProps } from '../Statistic';
@@ -10,7 +10,7 @@ import Operation from '../Operation';
 
 import './index.less';
 
-export type StatisticsCardProps = {
+export type StatisticCardProps = {
   /** 图表配置 */
   chart?: React.ReactNode;
   /** 数值统计配置 */
@@ -19,9 +19,12 @@ export type StatisticsCardProps = {
   chartPlacement?: 'right' | 'bottom' | 'left';
   /** 底部额外展示区域 */
   footer?: React.ReactNode;
-} & ProCardProps;
+} & CardProps;
 
-const StatisticsCard: React.FC<StatisticsCardProps> & {
+/** @deprecated */
+export type StatisticsCardProps = StatisticCardProps;
+
+const StatisticCard: React.FC<StatisticCardProps> & {
   Statistic: typeof Statistic;
   Divider: typeof Divider;
   Operation: typeof Operation;
@@ -65,22 +68,22 @@ const StatisticsCard: React.FC<StatisticsCardProps> & {
   const footerDom = footer && <div className={`${prefixCls}-footer`}>{footer}</div>;
 
   return (
-    <ProCard className={classString} {...others}>
+    <Card className={classString} {...others}>
       {contentDom}
       {children}
       {footerDom}
-    </ProCard>
+    </Card>
   );
 };
 
-const Group: React.FC<StatisticsCardProps> = (props) => (
-  <StatisticsCard bodyStyle={{ padding: 0 }} {...props} />
+const Group: React.FC<StatisticCardProps> = (props) => (
+  <StatisticCard bodyStyle={{ padding: 0 }} {...props} />
 );
 
-StatisticsCard.Statistic = Statistic;
-StatisticsCard.Divider = Divider;
-StatisticsCard.Operation = Operation;
-StatisticsCard.isProCard = true;
-StatisticsCard.Group = Group;
+StatisticCard.Statistic = Statistic;
+StatisticCard.Divider = Divider;
+StatisticCard.Operation = Operation;
+StatisticCard.isProCard = true;
+StatisticCard.Group = Group;
 
-export default StatisticsCard;
+export default StatisticCard;