+++ packages/table/src/utils/cellRenderToFromItem.tsx
@@ -138,10 +138,11 @@ function defaultRenderText<T>(config: {
         if (!columnProps?.renderFormItem) {
           return (
             <InlineErrorFormItem
+              errorType="popover"
               name={name}
               {...formItemProps}
               messageVariables={messageVariables}
-              initialValue={text || formItemProps?.initialValue}
+              initialValue={text !== undefined ? text : formItemProps?.initialValue}
             >
               {inputDom}
             </InlineErrorFormItem>
@@ -158,6 +159,7 @@ function defaultRenderText<T>(config: {
           {
             defaultRender: () => (
               <InlineErrorFormItem
+                errorType="popover"
                 name={name}
                 {...formItemProps}
                 messageVariables={messageVariables}
@@ -175,6 +177,7 @@ function defaultRenderText<T>(config: {
         );
         return (
           <InlineErrorFormItem
+            errorType="popover"
             name={spellNamePath(
               config.recordKey || config.index,
               columnProps?.key || columnProps?.dataIndex || config.index,